Adjust tile sizes, rename components
