+++ package.json
@@ -15,7 +15,7 @@
     "serve": "vite preview",
     "sprites": "npx -y @beyondtracks/spritezero-cli public/map/sprite public/map/icons/ --ratio=1 && npx -y @beyondtracks/spritezero-cli public/map/sprite@2x public/map/icons/ --ratio=2",
     "test": "npm run lint",
-    "tiles": "cd data && tippecanoe --base-zoom=6 --maximum-zoom=8 --low-detail=10 --full-detail=10 --no-tile-compression --maximum-tile-bytes=600000 --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --force -e ../public/map/tiles bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson && tippecanoe --base-zoom=9 --minimum-zoom=9 --maximum-zoom=11 --low-detail=10 --full-detail=10 --no-tile-compression --maximum-tile-bytes=600000 --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --allow-existing -e ../public/map/tiles stehende_gewaesser.geojson gewaessernetz.geojson bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson && tippecanoe --base-zoom=12 --minimum-zoom=12 --maximum-zoom=15 --full-detail=15 --no-tile-compression --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --allow-existing -e ../public/map/tiles stehende_gewaesser.geojson gewaessernetz.geojson bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson invekos_schlaege_2022_polygon.geojson && cd .."
+    "tiles": "cd data && tippecanoe --base-zoom=6 --maximum-zoom=8 --low-detail=10 --full-detail=10 --no-tile-compression --maximum-tile-bytes=600000 --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --force -e ../public/map/tiles bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson && tippecanoe --base-zoom=9 --minimum-zoom=9 --maximum-zoom=11 --low-detail=10 --full-detail=10 --no-tile-compression --maximum-tile-bytes=605000 --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --allow-existing -e ../public/map/tiles stehende_gewaesser.geojson gewaessernetz.geojson bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson && tippecanoe --base-zoom=12 --minimum-zoom=12 --maximum-zoom=15 --full-detail=15 --no-tile-compression --detect-shared-borders -x geo_id -x inspire_id -x geom_date_created --allow-existing -e ../public/map/tiles stehende_gewaesser.geojson gewaessernetz.geojson bdfl_l08_seehoehe_1200m.geojson bdfl_l13_besondere_lebensraumtypen.geojson bdfl_l16_grundwasserschutz_acker.geojson bdfl_l26_wasserrahmenrichtlinie_ertragslagen.geojson bdfl_l31_begruente_abflusswege.geojson bdfl_l32_belastete_gewaesser_polygon.geojson bdfl_l32_belastete_gewaesser_line.geojson bdfl_l33_pufferstreifen_belastete_gewaesser.geojson bdfl_l34_feucht_torfgebiete.geojson bdfl_l40_gab8.geojson invekos_schlaege_2022_polygon.geojson && cd .."
   },
   "dependencies": {
     "@mdi/font": "5.9.55",
