+++ src/components/TheSchlagInfoPanel.vue
