+++ src/components/TheTopicPanel.vue
@@ -111,7 +111,7 @@
                 class="pa-1"
               >
                 <div
-                  style="background: rgba(188, 143, 143, 0.3);"
+                  style="background-color: rgba(188, 143, 143, 0.3);"
                   class="colorBox"
                   :style="'opacity: ' + opacity + ';'"
                 />
